Imrpve mixin
