+++ tests/test_utils.py
@@ -1,5 +1,6 @@
 import logging
 import unittest
+from typing import Dict
 
 from django.http import HttpRequest
 from django.core.handlers.wsgi import WSGIRequest
@@ -49,3 +50,16 @@ class IsRequestInstanceTests(unittest.TestCase):
         """ rest_framework.request.Request """
         request = WSGIRequest({'REQUEST_METHOD': 'GET', 'wsgi.input': ''})
         self.assertTrue(utils.is_request_instance(request))
+
+
+class GetClientIpTests(unittest.TestCase):
+
+    def test_simple(self):
+        params: Dict[str, str] = {
+            'REQUEST_METHOD': 'GET',
+            'wsgi.input': '',
+            'REMOTE_ADDR': '127.0.0.0'
+        }
+        request = WSGIRequest(params)
+        ip = utils.get_client_ip(request)
+        assert ip == params['REMOTE_ADDR']
